Updated Java RMI test | client-to-server query
I decided to go ahead and create a MySQL instance on Amazon RDS. It was probably the cleanest solution. Honestly, if we're able to connect and manipulate the data no problem now while we're testing, we might as well stick with this implementation for the database. I already tested it with this Java RMI application I'm uploading and everything is smooth sailing so far. Currently the database connection credentials are hard-coded into the Java program, but I'll change that during this sprint.